fix(memory-v2): set 1h TTL on router cache, drop unused rawIds, hoist resolveUserName (#30177)
